Finished adding new functionality.
I still need to edit the comments and readme, but I can do that later.
Anyway, what I did was I made it so that you can only study new cards,
old cards, or all of them, because she just told us that there would
be none of the old pieces on the exam.